copyedit index.md
I couldn't tell from the quotation - do you mean to say the Hilbronner was trying to say they were superior to (better than) the consumer and other vendors, or that they were trying to convince consumers and other vendors that they were superior. I suspect it's the latter, but the "superior to'" construction makes it ambiguous.